Move level storage into separate package
